Report pdf fattura - Refactoring della struttura file e cartelle per il report fattura. Definito servizio InvoiceReportGenerator che si occupa della definizione del report e dei suoi dati.
