+++ src/Gdn.Web.Api.Vs/appsettings.json
@@ -6,6 +6,9 @@
     }
   },
   "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "SqlServerDefault": "Server=YOUR_SERVER;Initial Catalog=YOUR_DATABASE;Persist Security Info=False;User ID=YOUR_USERNAME;Password=YOUR_PWD;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+  },
   "CompanyData": {
     "Name": "Azienda di test",
     "Street": "Via di test, 20",
